+++ packages/cli/templates/webcomponents/igc-ts/date-time-input/default/index.ts
@@ -1,0 +1,19 @@
+import { TypeScriptFileUpdate } from "@igniteui/cli-core";
+import { IgniteUIForWebComponentsTemplate } from "../../../../../lib/templates/IgniteUIForWebComponentsTemplate";
+
+class IgcDateTimeInputTemplate extends IgniteUIForWebComponentsTemplate {
+	constructor() {
+		super(__dirname);
+		this.components = ["DateTime Input"];
+		this.controlGroup = "Data Entry & Display";
+		this.listInComponentTemplates = true;
+		this.id = "date-time-input";
+		this.projectType = "igc-ts";
+		this.name = "Date Time Input";
+		this.description = "basic IgcDateTimeInput";
+	}
+	protected addClassDeclaration(mainModule: TypeScriptFileUpdate, projPath: string, name: string, modulePath: string) {
+		// not applicable with custom module
+	}
+}
+module.exports = new IgcDateTimeInputTemplate();