+++ packages/cli/templates/webcomponents/igc-ts/projects/_base/files/rollup.config.mjs
@@ -1,11 +1,11 @@
-import nodeResolve from '@rollup/plugin-node-resolve';
 import babel from '@rollup/plugin-babel';
-import copy from 'rollup-plugin-copy-assets';
+import nodeResolve from '@rollup/plugin-node-resolve';
 import { rollupPluginHTML as html } from '@web/rollup-plugin-html';
 import { importMetaAssets } from '@web/rollup-plugin-import-meta-assets';
+import path from 'path';
+import copy from 'rollup-plugin-copy-assets';
 import { terser } from 'rollup-plugin-terser';
 import { generateSW } from 'rollup-plugin-workbox';
-import path from 'path';
 
 export default {
   input: 'index.html',