+++ packages/cli/templates/webcomponents/igc-ts/projects/_base/files/src/app/app.ts
@@ -1,6 +1,6 @@
-import { html, css, LitElement } from 'lit';
+import { Router } from '@vaadin/router';
+import { css, html, LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
-import { Router, } from '@vaadin/router';
 import { routes } from './app-routing.js';
 
 @customElement('app-root')
@@ -8,19 +8,20 @@ export class App extends LitElement {
   static styles = css`
     router-outlet {
       width: 100%;
-      height: 100%;
+    }
+
+    .outer-wrapper {
       display: flex;
-      text-align: center;
-      flex-flow: column nowrap;
-      justify-content: stretch;
-      align-items: center;
-      overflow: inherit;
+      justify-content: center;
+      height: 100%;
     }
   `;
 
   render() {
     return html`
-      <router-outlet></router-outlet>
+      <div class="outer-wrapper">
+        <router-outlet></router-outlet>
+      </div>
     `;
   }