+++ packages/cli/templates/webcomponents/igc-ts/projects/side-nav/files/src/app/app.ts
@@ -4,8 +4,6 @@ import { customElement } from 'lit/decorators.js';
 import {
   defineComponents,
   IgcNavDrawerComponent,
-  IgcNavDrawerItemComponent,
-  IgcNavDrawerHeaderItemComponent,
   IgcRippleComponent,
 } from 'igniteui-webcomponents';
 import { Router } from '@vaadin/router';
@@ -13,8 +11,6 @@ import { routes } from './app-routing.js';
 
 defineComponents(
   IgcNavDrawerComponent,
-  IgcNavDrawerItemComponent,
-  IgcNavDrawerHeaderItemComponent,
   IgcRippleComponent,
 );
 
@@ -22,7 +18,6 @@ defineComponents(
 export class App extends LitElement {
   static styles = css`
     router-outlet {
-      width: 100%;
       height: 100%;
       display: flex;
       text-align: center;
@@ -31,11 +26,15 @@ export class App extends LitElement {
       align-items: center;
       overflow: inherit;
     }
+
+    igc-nav-drawer {
+      float: left;
+    }
   `;
 
   render() {
     return html`
-      <igc-nav-drawer open=true>
+      <igc-nav-drawer open="true" position="relative">
         <igc-nav-drawer-header-item>Ignite UI CLI</igc-nav-drawer-header-item>
         ${routes.filter((element, index) => index < routes.length - 1).map(i => html`
           <igc-nav-drawer-item>