+++ packages/cli/templates/webcomponents/igc-ts/tabs/index.ts
@@ -1,0 +1,14 @@
+import { BaseComponent } from "@igniteui/cli-core";
+
+class IgcTabsComponent extends BaseComponent {
+	/**
+	 *
+	 */
+	constructor() {
+		super(__dirname);
+		this.name  = "Tabs";
+		this.group = "Layouts";
+		this.description = `Basic tabs component`;
+	}
+}
+module.exports = new IgcTabsComponent();