+++ packages/igx-templates/igx-ts/custom-templates/crm-grid/index.ts
@@ -41,7 +41,7 @@ class IgxGridCRMTemplate extends IgniteUIForAngularTemplate {
 			},
 			{ import: "FormsModule", from: "@angular/forms" }
 		];
-		this.packages = ["igniteui-angular-core@~13.1.0", "igniteui-angular-charts@~13.1.0"];
+		this.packages = ["igniteui-angular-core@~14.1.0", "igniteui-angular-charts@~14.1.0"];
 	}
 }
 module.exports = new IgxGridCRMTemplate();