+++ packages/igx-templates/igx-ts/financial-chart/default/index.ts
@@ -12,7 +12,7 @@ class IgxFinancialChartTemplate extends IgniteUIForAngularTemplate {
 		this.dependencies = [
 			{ import: "IgxFinancialChartModule", from: "igniteui-angular-charts" }
 		];
-		this.packages = ["igniteui-angular-core@~13.1.0", "igniteui-angular-charts@~13.1.0"];
+		this.packages = ["igniteui-angular-core@~14.1.0", "igniteui-angular-charts@~14.1.0"];
 	}
 }
 module.exports = new IgxFinancialChartTemplate();